+++ CHANGELOG.md
@@ -4,6 +4,7 @@ History of **Houston we have (no) problem** application with changes description
 
 ## Next release
 
+- `FEATURE` Doughnut health check chart
 - `FEATURE` Health check e-mail notification
 
 ## 0.3.0 (16.07.2025 - unpublished)